+++ src/components/header/index.tsx
@@ -1,10 +1,19 @@
 import React from 'react'
 import Image from 'next/image'
+import { useSession } from 'next-auth/client'
+
+import { SidebarMenu } from './sidebar-menu'
 
 export const Header = () => {
+  const [session] = useSession()
+
   return (
     <header>
-      <div className="flex justify-center p-4">
+      <div className="flex p-4 justify-between shadow-2xl">
+        {session && (
+          <SidebarMenu />
+        )}
+
         <Image
           src="/assets/svg/Header.svg"
           alt="Header"
@@ -13,6 +22,8 @@ export const Header = () => {
           placeholder="blur"
           blurDataURL={'/assets/svg/Header.svg'}
         />
+
+        <span></span>
       </div>
     </header>
   )